3.2-crud/stocks_products/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
django-filter==23.2